requirements.txt
asgiref==3.2.10
class-registry==2.1.2
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-tinymce==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
filters==1.3.2
filters-django==1.0.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
